chore: update next turbopack setup
