+++ packages/global/sdk/fastgpt-plugin.ts
@@ -1,1 +1,13 @@
-export * from '@fastgpt-sdk/plugin';
+export {
+  FastGPTPluginClient,
+  RunToolWithStream,
+  ToolDetailSchema,
+  ToolSimpleSchema,
+  ToolTagsNameMap
+} from '@fastgpt-sdk/plugin';
+export type {
+  AiproxyMapProviderType,
+  I18nStringStrictType,
+  ToolDetailType,
+  ToolSimpleType
+} from '@fastgpt-sdk/plugin';
