+++ packages/service/common/bullmq/index.ts
@@ -3,6 +3,7 @@ import {
   type Processor,
   Queue,
   type QueueOptions,
+  UnrecoverableError,
   Worker,
   type WorkerOptions
 } from 'bullmq';
@@ -150,4 +151,5 @@ export function getWorker<DataType, ReturnType = void>(
   return newWorker;
 }
 
-export * from 'bullmq';
+export { Queue, UnrecoverableError, Worker, delay };
+export type { ConnectionOptions, Job, Processor, QueueOptions, WorkerOptions } from 'bullmq';
