+++ packages/service/common/system/emptyModule.js
@@ -1,0 +1,14 @@
+const emptyModule = new Proxy(function emptyModule() {}, {
+  get() {
+    return emptyModule;
+  },
+  apply() {
+    return undefined;
+  },
+  construct() {
+    return {};
+  }
+});
+
+module.exports = emptyModule;
+module.exports.default = emptyModule;
