+++ packages/web/i18n/en/account_team.json
@@ -200,18 +200,6 @@
   "log_update_dataset": "【{{name}}】Updated [{{datasetType}}] named [{{datasetName}}]",
   "log_update_dataset_collaborator": "[{{name}}] Updated the collaborator named [{{datasetName}}] to: Organization: [{{orgList}}], Group: [{{groupList}}], Member [{{tmbList}}]; permissions updated to: [{{readPermission}}], [{{writePermission}}], [{{managePermission}}]",
   "log_update_publish_app": "【{{name}}】【{{operationName}}】【{{appType}}】 named [{{appName}}】",
-  "log_create_skill": "[{{name}}] Created a skill named [{{skillName}}]",
-  "log_create_skill_folder": "[{{name}}] Created a skill folder named [{{folderName}}]",
-  "log_delete_skill": "[{{name}}] Deleted the skill named [{{skillName}}]",
-  "log_deploy_skill": "[{{name}}] Deployed the skill named [{{skillName}}]",
-  "log_import_skill": "[{{name}}] Imported a skill named [{{skillName}}]",
-  "log_export_skill": "[{{name}}] Exported a skill named [{{skillName}}]",
-  "log_copy_skill": "[{{name}}] Copied a skill named [{{skillName}}]",
-  "log_move_skill": "[{{name}}] Moved a skill named [{{skillName}}] to [{{targetFolderName}}]",
-  "log_update_skill_collaborator": "[{{name}}] Updated collaborators of [{{skillType}}] [{{skillName}}] to: Organizations: [{{orgList}}], Groups: [{{groupList}}], Members: [{{tmbList}}]; Permission: [{{permission}}]",
-  "log_delete_skill_collaborator": "[{{name}}] Deleted collaborator [{{itemName}}]: [{{itemValueName}}] from [{{skillType}}] [{{skillName}}]",
-  "log_transfer_skill_ownership": "[{{name}}] Transferred ownership of skill [{{skillName}}] to [{{newOwnerName}}]",
-  "log_update_skill": "[{{name}}] Updated the skill named [{{skillName}}]",
   "log_user": "Operator",
   "login": "Log in",
   "manage_member": "Managing members",
@@ -250,12 +238,15 @@
   "search_test": "Search Test",
   "set_invoice_header": "Set up invoice header",
   "set_name_avatar": "Team avatar",
+  "skill.folder": "Folder",
+  "skill.skill": "Skill",
   "sync_immediately": "Synchronize now",
   "sync_member_failed": "Synchronization of members failed",
   "sync_member_success": "Synchronize members successfully",
   "total_team_members": "Total {{amount}} members",
   "transfer_app_ownership": "Transfer app ownership",
   "transfer_dataset_ownership": "Transfer dataset ownership",
+  "transfer_skill_ownership": "Transfer skill ownership",
   "transfer_ownership": "Transfer ownership",
   "transfer_team_ownership": "Transfer Team",
   "transfer_success": "Transfer successful",
@@ -277,21 +268,33 @@
   "update_dataset": "Update the knowledge base",
   "update_dataset_collaborator": "Knowledge Base Permission Changes",
   "update_publish_app": "Application update",
-  "create_skill": "Create skill",
-  "create_skill_folder": "Create skill folder",
-  "delete_skill": "Delete skill",
-  "deploy_skill": "Deploy skill",
-  "import_skill": "Import skill",
-  "export_skill": "Export skill",
-  "copy_skill": "Copy skill",
-  "move_skill": "Move skill",
   "update_skill_collaborator": "Update skill collaborator",
-  "delete_skill_collaborator": "Delete skill collaborator",
-  "transfer_skill_ownership": "Transfer skill ownership",
   "update_skill": "Update skill",
   "used_times_limit": "Limit",
   "user_name": "username",
   "user_team_invite_member": "Invite members",
   "user_team_leave_team": "Leave the team",
-  "user_team_leave_team_failed": "Failure to leave the team"
+  "user_team_leave_team_failed": "Failure to leave the team",
+  "waiting": "To be accepted",
+  "copy_skill": "Copy skill",
+  "create_skill": "Create skill",
+  "create_skill_folder": "Create skill folder",
+  "delete_skill": "Delete skill",
+  "delete_skill_collaborator": "Skill permission deletion",
+  "deploy_skill": "Deploy skill",
+  "export_skill": "Export skill",
+  "import_skill": "Import skill",
+  "log_copy_skill": "[{{name}}] Copied [{{skillType}}] named [{{skillName}}]",
+  "log_create_skill": "[{{name}}] Created [{{skillType}}] named [{{skillName}}]",
+  "log_create_skill_folder": "[{{name}}] Created a skill folder named [{{folderName}}]",
+  "log_delete_skill": "[{{name}}] Deleted [{{skillType}}] named [{{skillName}}]",
+  "log_delete_skill_collaborator": "[{{name}}] Deleted the [{{itemName}}] permission named [{{itemValueName}}] from [{{skillType}}] named [{{skillName}}]",
+  "log_deploy_skill": "[{{name}}] Deployed [{{skillType}}] named [{{skillName}}]",
+  "log_export_skill": "[{{name}}] Exported [{{skillType}}] named [{{skillName}}]",
+  "log_import_skill": "[{{name}}] Imported [{{skillType}}] named [{{skillName}}]",
+  "log_move_skill": "[{{name}}] Moved [{{skillType}}] named [{{skillName}}] to [{{targetFolderName}}]",
+  "log_transfer_skill_ownership": "[{{name}}] Transferred ownership of [{{skillType}}] named [{{skillName}}] from [{{oldOwnerName}}] to [{{newOwnerName}}]",
+  "log_update_skill": "[{{name}}] Updated [{{skillType}}] named [{{skillName}}]",
+  "log_update_skill_collaborator": "[{{name}}] Updated collaborators for [{{skillType}}] named [{{skillName}}] to: Organization: [{{orgList}}], Group: [{{groupList}}], Member [{{tmbList}}]; permissions updated to: [{{permission}}]",
+  "move_skill": "Move skill"
 }
