+++ packages/web/i18n/zh-CN/account_team.json
@@ -56,8 +56,6 @@
   "create_invitation_link": "创建邀请链接",
   "create_invoice": "开发票",
   "create_org": "创建部门",
-  "create_skill": "创建技能",
-  "create_skill_folder": "创建技能文件夹",
   "create_sub_org": "创建子部门",
   "dataset.api_file": "API 知识库",
   "dataset.common_dataset": "知识库",
@@ -66,8 +64,6 @@
   "dataset.folder_dataset": "文件夹",
   "dataset.website_dataset": "网站同步",
   "dataset.yuque_dataset": "语雀知识库",
-  "skill.folder": "技能文件夹",
-  "skill.skill": "技能",
   "delete": "删除",
   "delete_api_key": "删除api密钥",
   "delete_app": "删除工作台应用",
@@ -83,7 +79,6 @@
   "delete_from_team": "移出团队",
   "delete_group": "删除群组",
   "delete_org": "删除部门",
-  "delete_skill": "删除技能",
   "department": "部门",
   "edit_info": "编辑信息",
   "edit_member": "编辑用户",
@@ -163,8 +158,6 @@
   "log_create_group": "【{{name}}】创建了群组【{{groupName}}】",
   "log_create_invitation_link": "【{{name}}】创建了邀请链接【{{link}}】",
   "log_create_invoice": "【{{name}}】进行了开发票操作",
-  "log_create_skill": "【{{name}}】创建了名为【{{skillName}}】的技能",
-  "log_create_skill_folder": "【{{name}}】创建了名为【{{folderName}}】的技能文件夹",
   "log_delete_api_key": "【{{name}}】删除了名为【{{keyName}}】的api密钥",
   "log_delete_app": "【{{name}}】将名为【{{appName}}】的【{{appType}}】删除",
   "log_delete_app_collaborator": "【{{name}}】将名为【{{appName}}】的【{{appType}}】中名为【{{itemValueName}}】的【{{itemName}}】权限删除",
@@ -175,7 +168,6 @@
   "log_delete_dataset_collaborator": "【{{name}}】将名为【{{datasetName}}】的【{{datasetType}}】中名为【itemValueName】的【itemName】权限删除",
   "log_delete_department": "【{{name}}】删除了部门【{{departmentName}}】",
   "log_delete_evaluation": "【{{name}}】删除了名为【{{appName}}】的【{{appType}}】的评测数据",
-  "log_delete_skill": "【{{name}}】删除了名为【{{skillName}}】的技能",
   "log_delete_group": "【{{name}}】删除了群组【{{groupName}}】",
   "log_details": "详情",
   "log_export_app_chat_log": "【{{name}}】导出了名为【{{appName}}】的【{{appType}}】的聊天记录",
@@ -206,15 +198,6 @@
   "log_update_dataset": "【{{name}}】更新了名为【{{datasetName}}】的【{{datasetType}}】",
   "log_update_dataset_collaborator": "【{{name}}】将名为【{{datasetName}}】的【{{datasetType}}】的合作者更新为：组织：【{{orgList}}】，群组：【{{groupList}}】，成员【{{tmbList}}】；权限更新为：【{{readPermission}}】，【{{writePermission}}】，【{{managePermission}}】",
   "log_update_publish_app": "【{{name}}】【{{operationName}}】名为【{{appName}}】的【{{appType}}】",
-  "log_update_skill": "【{{name}}】更新了名为【{{skillName}}】的技能",
-  "log_deploy_skill": "【{{name}}】发布了名为【{{skillName}}】的技能",
-  "log_import_skill": "【{{name}}】导入了名为【{{skillName}}】的技能",
-  "log_export_skill": "【{{name}}】导出了名为【{{skillName}}】的技能",
-  "log_copy_skill": "【{{name}}】复制了名为【{{skillName}}】的技能",
-  "log_move_skill": "【{{name}}】将名为【{{skillName}}】的技能移动到【{{targetFolderName}}】",
-  "log_update_skill_collaborator": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】的协作者更新为：组织：【{{orgList}}】，群组：【{{groupList}}】，成员【{{tmbList}}】；权限更新为：【{{permission}}】",
-  "log_delete_skill_collaborator": "【{{name}}】删除了名为【{{skillName}}】的【{{skillType}}】的协作者【{{itemName}}】：【{{itemValueName}}】",
-  "log_transfer_skill_ownership": "【{{name}}】将名为【{{skillName}}】的技能所有权转移给了【{{newOwnerName}}】",
   "log_user": "操作人员",
   "login": "登录",
   "manage_member": "管理成员",
@@ -253,12 +236,15 @@
   "search_test": "搜索测试",
   "set_invoice_header": "设置发票抬头",
   "set_name_avatar": "团队头像 & 团队名",
+  "skill.folder": "文件夹",
+  "skill.skill": "技能",
   "sync_immediately": "立即同步",
   "sync_member_failed": "同步成员失败",
   "sync_member_success": "同步成员成功",
   "total_team_members": "共 {{amount}} 名成员",
   "transfer_app_ownership": "转移应用所有权",
   "transfer_dataset_ownership": "转移知识库所有权",
+  "transfer_skill_ownership": "转移技能所有权",
   "transfer_ownership": "转让所有者",
   "transfer_team_ownership": "转让团队",
   "transfer_success": "转让成功",
@@ -281,17 +267,32 @@
   "update_dataset_collaborator": "知识库权限更改",
   "update_publish_app": "应用更新",
   "update_skill": "更新技能",
-  "deploy_skill": "发布技能",
-  "import_skill": "导入技能",
-  "export_skill": "导出技能",
-  "copy_skill": "复制技能",
-  "move_skill": "移动技能",
   "update_skill_collaborator": "更新技能协作者",
-  "delete_skill_collaborator": "删除技能协作者",
-  "transfer_skill_ownership": "转移技能所有权",
   "used_times_limit": "有效人数",
   "user_name": "用户名",
   "user_team_invite_member": "邀请成员",
   "user_team_leave_team": "离开团队",
-  "user_team_leave_team_failed": "离开团队失败"
+  "user_team_leave_team_failed": "离开团队失败",
+  "waiting": "待接受",
+  "copy_skill": "复制技能",
+  "create_skill": "创建技能",
+  "create_skill_folder": "创建技能文件夹",
+  "delete_skill": "删除技能",
+  "delete_skill_collaborator": "技能权限删除",
+  "deploy_skill": "部署技能",
+  "export_skill": "导出技能",
+  "import_skill": "导入技能",
+  "log_copy_skill": "【{{name}}】复制了名为【{{skillName}}】的【{{skillType}}】",
+  "log_create_skill": "【{{name}}】创建了名为【{{skillName}}】的【{{skillType}}】",
+  "log_create_skill_folder": "【{{name}}】创建了名为【{{folderName}}】的技能文件夹",
+  "log_delete_skill": "【{{name}}】删除了名为【{{skillName}}】的【{{skillType}}】",
+  "log_delete_skill_collaborator": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】中名为【{{itemValueName}}】的【{{itemName}}】权限删除",
+  "log_deploy_skill": "【{{name}}】部署了名为【{{skillName}}】的【{{skillType}}】",
+  "log_export_skill": "【{{name}}】导出了名为【{{skillName}}】的【{{skillType}}】",
+  "log_import_skill": "【{{name}}】导入了名为【{{skillName}}】的【{{skillType}}】",
+  "log_move_skill": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】移动到【{{targetFolderName}}】",
+  "log_transfer_skill_ownership": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】的所有权从【{{oldOwnerName}}】转移到【{{newOwnerName}}】",
+  "log_update_skill": "【{{name}}】更新了名为【{{skillName}}】的【{{skillType}}】",
+  "log_update_skill_collaborator": "【{{name}}】将名为【{{skillName}}】的【{{skillType}}】的合作者更新为：组织：【{{orgList}}】，群组：【{{groupList}}】，成员【{{tmbList}}】；权限更新为：【{{permission}}】",
+  "move_skill": "移动技能"
 }
