+++ packages/web/i18n/zh-Hant/account_team.json
@@ -196,18 +196,6 @@
   "log_update_dataset": "【{{name}}】更新了名為【{{datasetName}}】的【{{datasetType}}】",
   "log_update_dataset_collaborator": "【{{name}}】將名為【{{datasetName}}】的【{{datasetType}}】的合作者更新為：組織：【{{orgList}}】，群組：【{{groupList}}】，成員【{{tmbList}}】；權限更新為：【{{readPermission}}】，【{{writePermission}}】，【{{managePermission}}】",
   "log_update_publish_app": "【{{name}}】【{{operationName}}】名為【{{appName}}】的【{{appType}}】",
-  "log_create_skill": "【{{name}}】建立了名為【{{skillName}}】的技能",
-  "log_create_skill_folder": "【{{name}}】建立了名為【{{folderName}}】的技能資料夾",
-  "log_delete_skill": "【{{name}}】刪除了名為【{{skillName}}】的技能",
-  "log_deploy_skill": "【{{name}}】發布了名為【{{skillName}}】的技能",
-  "log_import_skill": "【{{name}}】匯入了名為【{{skillName}}】的技能",
-  "log_export_skill": "【{{name}}】匯出了名為【{{skillName}}】的技能",
-  "log_copy_skill": "【{{name}}】複製了名為【{{skillName}}】的技能",
-  "log_move_skill": "【{{name}}】將名為【{{skillName}}】的技能移動到【{{targetFolderName}}】",
-  "log_update_skill_collaborator": "【{{name}}】將名為【{{skillName}}】的【{{skillType}}】的協作者更新為：組織：【{{orgList}}】，群組：【{{groupList}}】，成員【{{tmbList}}】；權限更新為：【{{permission}}】",
-  "log_delete_skill_collaborator": "【{{name}}】刪除了名為【{{skillName}}】的【{{skillType}}】的協作者【{{itemName}}】：【{{itemValueName}}】",
-  "log_transfer_skill_ownership": "【{{name}}】將名為【{{skillName}}】的技能所有權轉移給了【{{newOwnerName}}】",
-  "log_update_skill": "【{{name}}】更新了名為【{{skillName}}】的技能",
   "log_user": "操作人員",
   "login": "登入",
   "manage_member": "管理成員",
@@ -246,12 +234,15 @@
   "search_test": "搜索測試",
   "set_invoice_header": "設置發票抬頭",
   "set_name_avatar": "團隊頭像",
+  "skill.folder": "資料夾",
+  "skill.skill": "技能",
   "sync_immediately": "立即同步",
   "sync_member_failed": "同步成員失敗",
   "sync_member_success": "同步成員成功",
   "total_team_members": "共 {{amount}} 名成員",
   "transfer_app_ownership": "轉移應用程式所有權",
   "transfer_dataset_ownership": "轉移知識庫所有權",
+  "transfer_skill_ownership": "轉移技能所有權",
   "transfer_ownership": "轉讓所有者",
   "transfer_team_ownership": "轉讓團隊",
   "transfer_success": "轉讓成功",
@@ -272,22 +263,34 @@
   "update_data": "更新數據",
   "update_dataset": "更新知識庫",
   "update_dataset_collaborator": "知識庫權限更改",
-  "update_publish_app": "應用更新",
-  "create_skill": "建立技能",
-  "create_skill_folder": "建立技能資料夾",
-  "delete_skill": "刪除技能",
-  "deploy_skill": "發布技能",
-  "import_skill": "匯入技能",
-  "export_skill": "匯出技能",
-  "copy_skill": "複製技能",
-  "move_skill": "移動技能",
-  "update_skill_collaborator": "更新技能協作者",
-  "delete_skill_collaborator": "刪除技能協作者",
-  "transfer_skill_ownership": "轉移技能所有權",
   "update_skill": "更新技能",
+  "update_skill_collaborator": "技能權限更改",
+  "update_publish_app": "應用更新",
   "used_times_limit": "有效人數",
   "user_name": "使用者名稱",
   "user_team_invite_member": "邀請成員",
   "user_team_leave_team": "離開團隊",
-  "user_team_leave_team_failed": "離開團隊失敗"
+  "user_team_leave_team_failed": "離開團隊失敗",
+  "waiting": "待接受",
+  "copy_skill": "複製技能",
+  "create_skill": "建立技能",
+  "create_skill_folder": "建立技能資料夾",
+  "delete_skill": "刪除技能",
+  "delete_skill_collaborator": "技能權限刪除",
+  "deploy_skill": "部署技能",
+  "export_skill": "導出技能",
+  "import_skill": "導入技能",
+  "log_copy_skill": "【{{name}}】複製了名為【{{skillName}}】的【{{skillType}}】",
+  "log_create_skill": "【{{name}}】建立了名為【{{skillName}}】的【{{skillType}}】",
+  "log_create_skill_folder": "【{{name}}】建立了名為【{{folderName}}】的技能資料夾",
+  "log_delete_skill": "【{{name}}】刪除了名為【{{skillName}}】的【{{skillType}}】",
+  "log_delete_skill_collaborator": "【{{name}}】將名為【{{skillName}}】的【{{skillType}}】中名為【{{itemValueName}}】的【{{itemName}}】權限刪除",
+  "log_deploy_skill": "【{{name}}】部署了名為【{{skillName}}】的【{{skillType}}】",
+  "log_export_skill": "【{{name}}】導出了名為【{{skillName}}】的【{{skillType}}】",
+  "log_import_skill": "【{{name}}】導入了名為【{{skillName}}】的【{{skillType}}】",
+  "log_move_skill": "【{{name}}】將名為【{{skillName}}】的【{{skillType}}】移動到【{{targetFolderName}}】",
+  "log_transfer_skill_ownership": "【{{name}}】將名為【{{skillName}}】的【{{skillType}}】的所有權從【{{oldOwnerName}}】轉移到【{{newOwnerName}}】",
+  "log_update_skill": "【{{name}}】更新了名為【{{skillName}}】的【{{skillType}}】",
+  "log_update_skill_collaborator": "【{{name}}】將名為【{{skillName}}】的【{{skillType}}】的合作者更新為：組織：【{{orgList}}】，群組：【{{groupList}}】，成員【{{tmbList}}】；權限更新為：【{{permission}}】",
+  "move_skill": "移動技能"
 }
