+++ projects/app/package.json
@@ -4,9 +4,9 @@
   "private": false,
   "scripts": {
     "dev": "NODE_OPTIONS='--max-old-space-size=8192' npm run build:workers && next dev",
-    "dev:skill": "NODE_OPTIONS='--max-old-space-size=8192' npm run build:workers && tsx server.ts",
-    "dev:webpack": "NODE_OPTIONS='--max-old-space-size=8192' npm run build:workers && WEBPACK=1 next dev --webpack",
-    "build": "npm run build:workers && next build --debug --webpack",
+    "build": "npm run build:workers && next build --debug",
+    "build:webpack": "npm run build:workers && next build --webpack --debug",
+    "analyze": "next experimental-analyze",
     "start": "next start",
     "test": "vitest run -c vitest.config.ts",
     "test:watch": "vitest -c vitest.config.ts",
@@ -55,11 +55,13 @@
     "json5": "catalog:",
     "jsondiffpatch": "^0.7.2",
     "jsonwebtoken": "^9.0.2",
+    "katex": "0.16.22",
     "lodash": "catalog:",
     "jszip": "^3.10.1",
     "mermaid": "^10.9.4",
     "mime": "catalog:",
     "minio": "catalog:",
+    "mongoose": "^8.10.1",
     "nanoid": "^5.1.3",
     "next": "catalog:",
     "next-i18next": "catalog:",
@@ -106,7 +108,6 @@
     "@typescript-eslint/parser": "^6.21.0",
     "eslint": "catalog:",
     "eslint-config-next": "catalog:",
-    "next-rspack": "catalog:",
     "tsx": "^4.20.6",
     "typescript": "^5.1.3",
     "vitest": "^3.0.9"
