+++ projects/app/src/pageComponents/app/detail/Publish/Playground/index.tsx
@@ -16,6 +16,7 @@ import { ChatSidebarPaneEnum } from '@/pageComponents/chat/constants';
 import { useSystemStore } from '@/web/common/system/useSystemStore';
 
 const defaultPlaygroundVisibilityForm: PlaygroundVisibilityConfigType = {
+  showSkillReferences: true,
   showRunningStatus: true,
   showSkillReferences: false,
   showCite: true,
